Cleanup, finish support for StrVec
I've removed all the wrapping/saturating checks.
It's not clear to me what the right choice is in any of these scenarios, if there even is one. It should be up to ChillVec to either abort on growing too large or commit to UB in this library. In either case, the scenario where length is in danger of overflowing a usize is not the business of the other data structures.